src/code/requirements.txt
pdf2image
mysql-connector-python
pyzbar
pytz